Sync module states during non-fit
